refactor(rtm): extract AbstractResponseTranslator; unify placeholder stripping
CNR and IBS ResponseTranslator shared the whole translate()/findMatch()
pipeline: empty->"empty", "httperror|" splitting, static-template lookup with
{HTTPERROR} injection, invalid-template fallback, the two description-map
rewrite loops, and findMatch(). Introduce CNIC\AbstractResponseTranslator
holding that pipeline plus narrow abstract hooks the brands supply: the static
template container (templates()), the two description rewrite maps, the
human-readable field name (fieldName(): description vs message), and the
missing/empty required-field check that triggers the invalid fallback
(hasMissingRequiredFields(): CODE/DESCRIPTION vs status).

Resolve the latent placeholder-stripping inconsistency: CNR stripped unknown
{UPPER} tokens per-field via a callback (only inside DESCRIPTION), while IBS
stripped them globally across the whole response. Unify on the CNR per-field
callback as the single correct behaviour — global stripping risked removing
legitimate {UPPER} content from data fields. replacePlaceholders() is now shared
and parameterised by fieldName(). Add an IBS test locking in that a {TOKEN} in a
non-message data field is preserved.

{COMMAND}/$cmd handling folds into the shared findMatch() (harmless for IBS,
whose maps are empty), removing the former IBS UnusedParam suppression. No
behaviour change for existing CNR/IBS tests. Static methods use late static
binding so each brand's hooks resolve correctly. RSRMID-2893.
